requirements.txt
# Core dependencies
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pyyaml>=6.0
esprima>=4.0.1

# Progress and UI
tqdm>=4.65.0

# Template engine
jinja2>=3.1.2

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1

# Code quality
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0